Added Falcon Punch option in configuration UI.  When checked this enables the following:
The "increment_pulse_because_birds_of_prey_are_douchebags" function is triggered by the "failedFalconPunch" command, which is broadcasted whenever a state pulses.
This function checks to see if the state's next pulse is within 100 milliseconds, and will then increment the next pulse by 100 milliseconds plus a random seed.

This should prevent actions on different sessions from showing up on the same server tick, which could be picked up by heuristics targeted at identifying multiboxers.  This may severely impact performance, which is why this is optional at this time.  On or before January 1st this will be made mandatory, so testing is encouraged.